Static code analysis (part 1)
Some issues of the below types have been addressed that were identified using the Visual Studio static code analysis. Only issues that are easy and safe to fix are included in this set.

- initialised variables
- removed unused variables
- removed unreachable code
- renaming to avoid ‘hidden’ variables
- prevent overflow in loop where decrementing UINT32 and checking if > 0
- avoid unintentional possible dataloss by assigning large variable to small variable (change type or add explicit cast)

+++ dynadjust/dynadjust/dnaimport/dnainterop.hpp
@@ -222,7 +222,7 @@ private:
 	string ParseAngularValue(const string& sBuf, const string& calling_function);
 	string ParseLinearValue(const string& sBuf, const string& msrName, const string& calling_function);
 	string ParseGPSMsrValue(const string& sBuf, const string& element, const string& calling_function);
-	string ParseGPSVarValue(const string& sBuf, const string& element, const UINT16 location, const UINT16 width, const string& calling_function);
+	string ParseGPSVarValue(const string& sBuf, const string& element, const UINT32 location, const UINT32 width, const string& calling_function);
 	string ParseInstrumentValue(const string& sBuf, const string& calling_function);
 	string ParseTargetValue(const string& sBuf, const string& calling_function);
 	string ParseTarget2Value(const string& sBuf, const string& calling_function);
